Ajustado as telas de departamento e de calendario para o usuario lider

+++ app/Livewire/TableDpt.php
@@ -7,6 +7,9 @@ use Illuminate\Http\Request;
 use App\Models\Departamento;
 use App\Models\DepartamentoUsuario;
 use App\Models\User;
+use App\Models\Evento;
+use App\Models\lider;
+use App\Models\Pedido;
 use Livewire\WithPagination;
 use Illuminate\Support\Facades\DB;
 
@@ -26,11 +29,27 @@ class TableDpt extends Component
 
     public function render(){
         //Pegando todos os departamentos pela busca
-        if($this->busca){
-            $allDpts = Departamento::where('nome','like','%'.$this->busca.'%')->paginate(4);
-        }else{
-            $allDpts = Departamento::paginate(4);
+        if(auth()->user()->tipo === 'pastor'){
+            if($this->busca){
+                $allDpts = Departamento::where('nome','like','%'.$this->busca.'%')->paginate(4);
+            }else{
+                $allDpts = Departamento::paginate(4);
+            }
+        }elseif(auth()->user()->tipo === 'lider'){
+            if($this->busca){
+                $allDpts = lider::join('departamentos', 'departamentos.id', '=', 'lideres.departamento_id')
+                ->select('lideres.*', 'departamentos.*')
+                ->where('departamento.nome','like','%'.$this->busca.'%')
+                ->where('lideres.user_id',auth()->user()->id)
+                ->paginate(4);
+            }else{
+                $allDpts = lider::join('departamentos', 'departamentos.id', '=', 'lideres.departamento_id')
+                ->select('lideres.*', 'departamentos.*')
+                ->where('lideres.user_id',auth()->user()->id)
+                ->paginate(4);
+            }
         }
+
 
 
         for($x = 0;$x<count($allDpts);$x++){
@@ -68,7 +87,10 @@ class TableDpt extends Component
     public function deleteDpt($idDpt){
         //dd($idDpt);
 
-        if(Departamento::where('id',$idDpt)->exists() && DepartamentoUsuario::where('departamento_id',$idDpt)->exists()){
+        if(Departamento::where('id',$idDpt)->exists() && DepartamentoUsuario::where('departamento_id',$idDpt)->exists() && Evento::where('departamento_id',$idDpt)->exists() && lider::where('departamento_id',$idDpt)->exists() && Pedido::where('departamento_id',$idDpt)->exists()){
+            Pedido::where('departamento_id',$idDpt)->delete();
+            Evento::where('departamento_id',$idDpt)->delete();
+            lider::where('departamento_id',$idDpt)->delete();
             DepartamentoUsuario::where('departamento_id',$idDpt)->delete();
             Departamento::where('id',$idDpt)->delete();
